% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{comparison}
\alias{comparison}
\title{comparison}
\usage{
comparison(graph, graphRandom, method1 = c("walktrap", "edgeBetweenness",
  "fastGreedy", "leadingEigen", "louvain", "spinglass", "labelProp",
  "infomap", "optimal"), method2 = c("walktrap", "edgeBetweenness",
  "fastGreedy", "leadingEigen", "louvain", "spinglass", "labelProp",
  "infomap", "optimal"), type = c("dependent", "independent"),
  directed = FALSE, weights = NULL, steps = 4, spins = 25,
  e.weights = NULL, v.weights = NULL, nb.trials = 10)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{graphRandom}{The output of random function.}

\item{method1}{The first clustering method, one of "walktrap", 
"edgeBetweenness", "fastGreedy", "louvain", "spinglass", "leadingEigen",
"labelProp", "infomap","optimal".}

\item{method2}{The second custering method one of "walktrap",
"edgeBetweenness","fastGreedy", "louvain", "spinglass", "leadingEigen",
"labelProp", "infomap","optimal".}

\item{type}{The type of robin costruction dependent or independent.}

\item{directed}{This argument is settable only for "edgeBetweenness" method.}

\item{weights}{This argument is not settable for "infomap" method.}

\item{steps}{This argument is settable only for "leadingEigen"and"walktrap" 
method.}

\item{spins}{This argument is settable only for "infomap" method.}

\item{e.weights}{This argument is settable only for "infomap" method.}

\item{v.weights}{This argument is settable only for "infomap" method.}

\item{nb.trials}{This argument is settable only for "infomap" method.}
}
\value{
A list object
}
\description{
A procedure to compare two different methods of community 
detection.
}
\examples{
Comp <- comparison(graph=graph, graphRandom=graphRandom, 
method1="louvain", method2="fastGreedy",type="independent")
}
