% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptkFunctions.R
\name{gpLogLikeGradients}
\alias{gpLogLikeGradients}
\title{gpLogLikeGradients}
\usage{
gpLogLikeGradients(
  model,
  X = model$X,
  M,
  X_u,
  gX_u.return = FALSE,
  gX.return = FALSE,
  g_beta.return = FALSE
)
}
\arguments{
\item{model}{}

\item{X}{}

\item{M}{}

\item{X_u}{}

\item{gX_u.return}{}

\item{gX.return}{}

\item{g_beta.return}{}
}
\value{

}
\description{
gpLogLikeGradients
}
