% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{intl}
\alias{intl}
\title{Primitive object of internal function}
\usage{
intl(nm)
}
\arguments{
\item{nm}{The name of the internal function, as a string or symbol.}
}
\value{
A primitive object.
}
\description{
You normally don't see internal primitives in the wild as they stay
hidden in the bushes and only go out at night. `intl()` grabs them
for you.
}
\examples{
intl("eval")
prim_ptr(intl("eval"))
}
