% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptkFunctions.R
\name{localCovarianceGradients}
\alias{localCovarianceGradients}
\title{localCovarianceGradients}
\usage{
localCovarianceGradients(model, y, dimension)
}
\arguments{
\item{model}{}

\item{y}{}

\item{dimension}{}
}
\value{

}
\description{
localCovarianceGradients
}
