% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{plotComm}
\alias{plotComm}
\title{plotComm}
\usage{
plotComm(graph, members)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{members}{A membership vector of the community structure, the output of
membershipCommunities.}
}
\value{
Creates an interactive plot with colorful communities, a D3 
JavaScript network graph.
}
\description{
Graphical interactive representation of the network and its 
communities.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
members <- membershipCommunities (graph=graph, method="louvain")
plotComm(graph, members)
}
