% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotCommu}
\alias{plotCommu}
\title{plotCommu}
\usage{
plotCommu(graph, method)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{method}{The clustering method, one of "walktrap", "edgeBetweenness", 
"fastGreedy", "louvain", "spinglass", "leadingEigen", "labelProp", "infomap",
"optimal".}
}
\value{
An interactive plot.
}
\description{
An interactive 3D plot of the communities.
}
\examples{
plotCommu(graph=graph, method="louvain")
}
