% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotRobin}
\alias{plotRobin}
\title{plotRobin}
\usage{
plotRobin(graph, model = Proc$viMean, modelR = Proc$viMeanRandom,
  legend = c("real data", "null model"))
}
\arguments{
\item{graph}{The output of prepGraph}

\item{model}{The viMean output of the robinProc function.}

\item{modelR}{The viMeanRandom output of the robinProc function.}

\item{legend}{The legend for the graph. The default is c("real data", 
"null model").}
}
\value{
A plot.
}
\description{
The plot of the two VI curves, the VI of the null model and 
of the real graph.
}
\examples{
plotRobin(graph=graph,model=Proc$viMean,modelR=Proc$viMeanRandom, 
legend=c("real data", "null model"))
}
