% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotRobinComp}
\alias{plotRobinComp}
\title{plotRobinComp}
\usage{
plotRobinComp(graph, model1, modelR1, model2, modelR2, measure = c("vi",
  "nmi", "split.join", "adjusted.rand"), legend = c("real data",
  "null model"), legend1vs2 = c("method1", "method2"),
  title1 = "Method1", title2 = "Method2",
  title1vs2 = "Method1 vs Method2")
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{model1}{The Mean1 output of the comparison function.}

\item{modelR1}{The MeanRandom1 output of the comparison function.}

\item{model2}{The Mean2 output of the comparison function.}

\item{modelR2}{The MeanRandom2 output of the comparison function.}

\item{measure}{The stability measure "vi", "nmi", "split.join", 
"adjusted.rand".}

\item{legend}{The legend for the graph. The default is c("real data", 
"null model").}

\item{legend1vs2}{The legend for the two methods.
The default is c("method1","method2")}

\item{title1}{The title for the plot of the first method. 
The default is "Method1".}

\item{title2}{The title for the plot of the second method.
The default is "Method2".}

\item{title1vs2}{The title for the plot combined. The default is "Method1 vs
Method2".}
}
\value{
A ggplot object in a grid format.
}
\description{
plotRobinComp
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
graphRandom <- random(graph=graph)
Comp <- robinComp(graph=graph, graphRandom=graphRandom, method1="louvain", 
method2="fastGreedy", measure="vi", type="independent") 
plotRobinComp(graph=graph, model1=Comp$Mean1, 
model2=Comp$Mean2,modelR1=Comp$MeanRandom1, modelR2=Comp$MeanRandom2,
measure="vi", legend=c("real data", "null model"),
legend1vs2=c("Louvain", "Fast Greedy"), title1="Louvain",
title2="Fast Greedy", title1vs2="Louvain vs Fast Greedy")

}
