% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotRobinCompare}
\alias{plotRobinCompare}
\title{plotRobinCompare}
\usage{
plotRobinCompare(graph, model1, modelR1, model2, modelR2,
  legend = c("real data", "null model"), legend1vs2 = c("method1",
  "method2"), title1 = "Method1", title2 = "Method2",
  title1vs2 = "Method1 vs Method2")
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{model1}{The viMean1 output of the comparison function.}

\item{modelR1}{The viMeanRandom1 output of the comparison function.}

\item{model2}{The viMean2 output of the comparison function.}

\item{modelR2}{The viMeanRandom2 output of the comparison function.}

\item{legend}{The legend for the graph. The default is c("real data", 
"null model").}

\item{legend1vs2}{The legend for the two methods.
The default is c("method1","method2")}

\item{title1}{The title for the plot of the first method. 
The default is "Method1".}

\item{title2}{The title for the plot of the second method.
The default is "Method2".}

\item{title1vs2}{The title for the plot combined. The default is "Method1 vs
Method2".}
}
\value{
A plot.
}
\description{
plotRobinCompare
}
\examples{
plotRobinCompare(graph=graph, model1=Comp$viMean1, 
model2=Comp$viMean2,modelR1=Comp$viMeanRandom1, modelR2=Comp$viMeanRandom2,
legend=c("real data", "null model"),legend1vs2=c("Louvain", "Fast Greedy"),
title1="Louvain",title2="Fast Greedy",
title1vs2="Louvain vs Fast Greedy")

}
