% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{prim}
\alias{prim}
\title{Primitive object}
\usage{
prim(nm)
}
\arguments{
\item{nm}{The name of an exported primitive as a string or symbol.}
}
\description{
This just grabs a primitive object (type BUILTINSXP or SPECIALSXP)
from the base environment.
}
