% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{prim_invoke}
\alias{prim_invoke}
\alias{prim_ptr_invoke}
\title{Invoke a primitive function}
\usage{
prim_invoke(prim, call, args = NULL, rho = caller_env())

prim_ptr_invoke(ptr, call, args = NULL, rho = caller_env())
}
\arguments{
\item{prim}{A primitive function.}

\item{call}{The quoted call that the primitive will think it was
invoked with.}

\item{args}{A pairlist of arguments.}

\item{rho}{The environment that the primitive will think it was
invoked in.}

\item{ptr}{A pointer object to a primitive function.}
}
\description{
Invoke a primitive function
}
\examples{
prim_invoke(list, NULL, pairlist(1, "a"))
}
