% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{prim_offset}
\alias{prim_offset}
\title{Offset of primitive function}
\usage{
prim_offset(prim)
}
\arguments{
\item{prim}{A primitive function.}
}
\description{
This returns the offset of a primitive function in `R_FunTab`.
}
\examples{
prim_offset(base::c)
prim_offset(base::list)
}
