% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{prim_ptr}
\alias{prim_ptr}
\title{Function pointer of primitive function}
\usage{
prim_ptr(prim)
}
\arguments{
\item{prim}{A primitive function.}
}
\value{
A list containing a pointer object to the C implementation
  of the primitive function.
}
\description{
Function pointer of primitive function
}
\examples{
prim_ptr(base::c)
prim_ptr(base::list)
}
