% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{random}
\alias{random}
\title{random}
\usage{
random(graph)
}
\arguments{
\item{graph}{The output of prepGraph.}
}
\value{
A randomly rewired graph.
}
\description{
Randomly rewire the edges while preserving the original graph's 
degree distribution.
}
\examples{
graphRandom <- random(graph=graph)
}
