% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{randomNoW}
\alias{randomNoW}
\title{randomNoW}
\usage{
randomNoW(graph, verbose = FALSE)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{verbose}{flag for verbose output (default as FALSE)}
}
\value{
An igraph object, a randomly rewired graph.
}
\description{
This function randomly rewires the edges while preserving the original graph's 
degree distribution.
}
\keyword{internal}
