% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBINWeighted.R
\name{randomWeight}
\alias{randomWeight}
\title{randomWeight}
\usage{
randomWeight(graph, rewire.w.type = "Rewire", verbose = FALSE)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{rewire.w.type}{method to rewire weighted graphs one of "Rewire",
"Shuffle","Garlaschelli","Sum". "Garlaschelli" method only for count weights,
"Sum" method only for continuous weights.}

\item{verbose}{flag for verbose output (default as FALSE)}
}
\value{
An igraph object, a randomly rewired graph.
}
\description{
This function randomly rewires the edges while preserving the original graph's
degree distribution.
}
\keyword{internal}
