% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{rewireCompl}
\alias{rewireCompl}
\title{rewireCompl}
\usage{
rewireCompl(
  data,
  number,
  community,
  method = c("walktrap", "edgeBetweenness", "fastGreedy", "louvain", "spinglass",
    "leadingEigen", "labelProp", "infomap", "optimal", "leiden", "other"),
  ...,
  measure = c("vi", "nmi", "split.join", "adjusted.rand"),
  FUN = NULL
)
}
\arguments{
\item{data}{The output of prepGraph}

\item{number}{Number of rewiring trials to perform.}

\item{community}{Community to compare with.}

\item{method}{The clustering method, one of "walktrap", "edgeBetweenness", 
"fastGreedy", "louvain", "spinglass", "leadingEigen", "labelProp", "infomap".}

\item{...}{additional parameters to use with any of the previous described 
methods (see igraph package community detection methods for more details 
i.e. \link[igraph]{cluster_walktrap})}

\item{measure}{The measure for the comparison of the communities "vi", "nmi",
"split.join", "adjusted.rand"}

\item{FUN}{see \code{\link{methodCommunity}}.}
}
\description{
rewires the graph, creates the communities and 
compares the communities through different measures.
}
\keyword{internal}
