% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{robinAUCTest}
\alias{robinAUCTest}
\title{robinAUCTest}
\usage{
robinAUCTest(graph, model1, model2)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{model1}{The viMean output of the robinProc function (or the viMean1 
output of the comparison function).}

\item{model2}{The viMeanRandom output of the robinProc function (or the 
viMean2 output of the comparison function).}
}
\value{
A list
}
\description{
Calculate the area under the two curves with a spline approach.
}
\examples{
robinAUCTest(graph=graph,model1=Proc$viMean,
model2=Proc$viMeanRandom)
}
