% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{robinCompare}
\alias{robinCompare}
\title{robinCompare}
\usage{
robinCompare(
  graph,
  method1 = c("walktrap", "edgeBetweenness", "fastGreedy", "leadingEigen", "louvain",
    "spinglass", "labelProp", "infomap", "optimal", "leiden", "other"),
  args1 = list(),
  method2 = c("walktrap", "edgeBetweenness", "fastGreedy", "leadingEigen", "louvain",
    "spinglass", "labelProp", "infomap", "optimal", "leiden", "other"),
  args2 = list(),
  FUN1 = NULL,
  FUN2 = NULL,
  measure = c("vi", "nmi", "split.join", "adjusted.rand"),
  type = c("independent", "dependent"),
  verbose = TRUE
)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{method1}{The first clustering method, one of "walktrap", 
"edgeBetweenness", "fastGreedy", "louvain", "spinglass", "leadingEigen",
"labelProp", "infomap","optimal".}

\item{method2}{The second custering method one of "walktrap",
"edgeBetweenness","fastGreedy", "louvain", "spinglass", "leadingEigen",
"labelProp", "infomap","optimal".}

\item{FUN1}{personal designed function when method1 is "others". 
see \code{\link{methodCommunity}}.}

\item{FUN2}{personal designed function when method2 is "others". 
see \code{\link{methodCommunity}}.}

\item{measure}{The stability measure, one of "vi", "nmi", "split.join", 
"adjusted.rand" all normalized and used as distances.
"nmi" refers to 1- nmi and "adjusted.ran" refers to 1-adjusted.rand.}

\item{type}{The type of robin construction, dependent or independent.}

\item{verbose}{flag for verbose output (default as TRUE).}

\item{weights}{This argument is not settable for "infomap" method.}

\item{steps}{This argument is settable only for "leadingEigen"and"walktrap" 
method.}

\item{spins}{This argument is settable only for "infomap" method.}

\item{e.weights}{This argument is settable only for "infomap" method.}

\item{v.weights}{This argument is settable only for "infomap" method.}

\item{nb.trials}{This argument is settable only for "infomap" method.}

\item{directed}{This argument is settable only for "edgeBetweenness" method.}

\item{objective_function}{Whether to use the Constant Potts Model (CPM) or 
modularity. Must be either "CPM" or "modularity".}

\item{n_iterations}{the number of iterations to iterate the Leiden algorithm. 
Each iteration may improve the partition further.This argument is settable 
only for "leiden".}

\item{resolution}{only for "louvain" and "leiden". Optional resolution 
parameter, lower values typically yield fewer, larger clusters (default=1).}
}
\value{
A list object with two matrices:
- the matrix "Mean1" with the means of the procedure for the first method 
- the matrix "Mean2" with the means of the procedure for the second method
}
\description{
This function compares the robustness of two community 
detection algorithms.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
robinCompare(graph=graph, method1="louvain", args1 = list(resolution=0.8),
            method2="leiden", args2=list(objective_function ="modularity"), 
            measure="vi", type="independent")
}
