% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{robinProc}
\alias{robinProc}
\title{robinProc}
\usage{
robinProc(graph, graphRandom, method, type = c("dependent",
  "independent"), directed = FALSE, weights = NULL, steps = 4,
  spins = 25, e.weights = NULL, v.weights = NULL, nb.trials = 10)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{graphRandom}{The output of random function.}

\item{method}{The clustering method, one of "walktrap", "edgeBetweenness", 
"fastGreedy", "louvain", "spinglass", "leadingEigen", "labelProp", "infomap",
"optimal".}

\item{type}{The type of robin costruction dependent or independent data}

\item{directed}{This argument is settable only for "edgeBetweenness" method.}

\item{weights}{this argument is not settable for "infomap" method.}

\item{steps}{this argument is settable only for "leadingEigen"and"walktrap" 
method.}

\item{spins}{This argument is settable only for "infomap" method.}

\item{e.weights}{This argument is settable only for "infomap" method.}

\item{v.weights}{This argument is settable only for "infomap" method.}

\item{nb.trials}{This argument is settable only for "infomap" method.}
}
\value{
A list object.
}
\description{
A procedure to examine the stability of the partition recovered 
against random perturbations of the original graph structure.
}
\examples{
Proc<- robinProc(graph=graph,graphRandom=graphRandom, 
method="louvain",type="independent")
}
