% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBINFast.R
\name{robinRobustFast}
\alias{robinRobustFast}
\title{robinRobustFast}
\usage{
robinRobustFast(
  graph,
  graphRandom,
  method = c("walktrap", "edgeBetweenness", "fastGreedy", "louvain", "spinglass",
    "leadingEigen", "labelProp", "infomap", "optimal", "leiden", "other"),
  ...,
  FUN = NULL,
  measure = c("vi", "nmi", "split.join", "adjusted.rand"),
  verbose = TRUE
)
}
\description{
robinRobustFast
}
