% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{sym_table}
\alias{sym_table}
\title{Symbol table}
\usage{
sym_table()
}
\value{
A list of 4019 elements. Each element is a pairlist of
  symbols.
}
\description{
This returns the hash buckets of R's symbol table.
}
