% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createITPSplineResult}
\alias{createITPSplineResult}
\title{createITPSplineResult}
\usage{
createITPSplineResult(graph, model1, model2, measure = c("vi", "nmi",
  "split.join", "adjusted.rand"), muParam = 0, orderParam = 4,
  nKnots = 7, BParam = 10000, isPaired = TRUE)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{model1}{The Mean output of the robinProc function (or the Mean1 
output of the comparison function).}

\item{model2}{The MeanRandom output of the robinProc function (or the 
Mean2 output of the comparison function).}

\item{measure}{The measure for the comparison of the communities "vi", "nmi",
"split.join", "adjusted.rand".}

\item{muParam}{the mu parameter for ITP2bspline (default 0).}

\item{orderParam}{the order parameter for ITP2bspline (default 4).}

\item{nKnots}{the nknots parameter for ITP2bspline (default 7).}

\item{BParam}{the B parameter for ITP2bspline (default 10000).}

\item{isPaired}{the paired parameter for ITP2bspline (default TRUE).}
}
\value{
an ITP2 object
}
\description{
creates an fdatest::ITP2 class object
}
