% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{plotGraph}
\alias{plotGraph}
\title{plotGraph}
\usage{
plotGraph(graph)
}
\arguments{
\item{graph}{The output of prepGraph.}
}
\value{
Creates an interactive plot, a D3 JavaScript network graph.
}
\description{
Graphical interactive representation of the network.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
plotGraph (graph)
}
