% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotRobin}
\alias{plotRobin}
\title{plotRobin}
\usage{
plotRobin(graph, model, modelR, measure = c("vi", "nmi", "split.join",
  "adjusted.rand"), legend = c("real data", "null model"))
}
\arguments{
\item{graph}{The output of prepGraph}

\item{model}{The viMean output of the robinProc function.}

\item{modelR}{The viMeanRandom output of the robinProc function.}

\item{measure}{The stability measure "vi", "nmi", "split.join", 
"adjusted.rand".}

\item{legend}{The legend for the graph. The default is c("real data", 
"null model").}
}
\value{
A ggplot object.
}
\description{
The plot of the two curves, the measures of the null model and 
of the real graph.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
graphRandom <- random(graph=graph)
Proc <- robin(graph=graph, graphRandom=graphRandom, method="louvain",
type="independent")
plotRobin(graph=graph, model=Proc$MeanReal, modelR=Proc$MeanRandom,
measure="vi", legend=c("real data", "null model"))
}
