% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN_plots.R
\name{plotRobin}
\alias{plotRobin}
\title{plotRobin}
\usage{
plotRobin(
  graph,
  model1,
  model2,
  legend = c("model1", "model2"),
  title = "Robin plot"
)
}
\arguments{
\item{graph}{The output of prepGraph}

\item{model1}{The Mean output of the robinRobust function or the Mean1 
output of robinCompare.}

\item{model2}{The MeanRandom output of the robinRobust function or the Mean2 
output of robinCompare.}

\item{legend}{The legend for the graph. The default is c("model1", 
"model2"). If using robinRobust is recommended c("real data", "null model"), 
if using robinCompare, enter the names of the community detection 
algorithms.}

\item{title}{The title for the graph. The default is "Robin plot".}
}
\value{
A ggplot object.
}
\description{
This function plots two curves: the measure of the null model and the measure
of the real graph or the measure of the two community detection algorithms.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
graphRandom <- random(graph=graph)
Proc <- robinRobust(graph=graph, graphRandom=graphRandom, method="louvain",
measure="vi", type="independent")
plotRobin(graph=graph, model1=Proc$Mean, model2=Proc$MeanRandom
, legend=c("real data", "null model"))

}
