% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN_external_function.R
\name{random}
\alias{random}
\title{random}
\usage{
random(graph, rewire.w.type = "Rewire", verbose = FALSE)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{rewire.w.type}{for weighted graph. Option to rewire one of "Rewire",
"Shuffle","Garlaschelli","Sum". "Garlaschelli" method only for count weights,
"Sum" method only for continuous weights.}

\item{verbose}{flag for verbose output (default as FALSE)}
}
\value{
An igraph object, a randomly rewired graph.
}
\description{
This function randomly rewires the edges while preserving the original graph's 
degree distribution.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
graphRandom <- random(graph=graph)
}
