% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBINWeighted.R
\name{randomWeight}
\alias{randomWeight}
\title{randomWeight}
\usage{
randomWeight(graph, distribution = "NegBinom", verbose = FALSE)
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{verbose}{flag for verbose output (default as FALSE)}
}
\value{
An igraph object, a randomly rewired graph.
}
\description{
This function randomly rewires the edges while preserving the original graph's
degree distribution.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
graphRandom <- randomWeight(graph=graph)
}
