% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBINWeighted.R
\name{rewireComplWeight}
\alias{rewireComplWeight}
\title{rewireComplWeight}
\usage{
rewireComplWeight(
  data,
  number,
  community,
  method = c("walktrap", "edgeBetweenness", "fastGreedy", "louvain", "spinglass",
    "leadingEigen", "labelProp", "infomap", "optimal", "other"),
  FUN = NULL,
  measure = c("vi", "nmi", "split.join", "adjusted.rand"),
  directed = FALSE,
  weights = NULL,
  steps = 4,
  spins = 25,
  e.weights = NULL,
  v.weights = NULL,
  nb.trials = 10,
  resolution = 1,
  n_iterations = 2,
  objective_function = c("CPM", "modularity"),
  distrib = "NegBinom"
)
}
\arguments{
\item{data}{The output of prepGraph.}

\item{number}{Number of rewiring trials to perform.}

\item{community}{Community to compare with.}

\item{method}{The clustering method, one of "walktrap", "edgeBetweenness",
"fastGreedy", "louvain", "spinglass", "leadingEigen", "labelProp", "infomap".}

\item{FUN}{see \code{\link{methodCommunity}}.}

\item{measure}{The measure for the comparison of the communities "vi", "nmi",
"split.join", "adjusted.rand".}

\item{directed}{This argument is settable only for "edgeBetweenness" method}

\item{weights}{this argument is not settable for "infomap" method.}

\item{steps}{this argument is settable only for "leadingEigen"and"walktrap"
method.}

\item{spins}{This argument is settable only for "infomap" method.}

\item{e.weights}{This argument is settable only for "infomap" method.}

\item{v.weights}{This argument is settable only for "infomap" method.}

\item{nb.trials}{This argument is settable only for "infomap" method.}

\item{resolution}{only for "louvain" and "leiden". Optional resolution 
parameter, lower values typically yield fewer, larger clusters (default=1).}

\item{n_iterations}{the number of iterations to iterate the Leiden algorithm. 
Each iteration may improve the partition further.This argument is settable 
only for "leiden".}

\item{objective_function}{Whether to use the Constant Potts Model (CPM) or 
modularity. Must be either "CPM" or "modularity".}
}
\description{
rewires the weighted graph, creates the communities and
compares the communities through different measures.
}
\keyword{internal}
