% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rewireOnl}
\alias{rewireOnl}
\title{rewireOnl}
\usage{
rewireOnl(data, number)
}
\arguments{
\item{data}{The output of prepGraph}

\item{number}{Number of rewiring trials to perform.}
}
\value{
An igraph object
}
\description{
makes the rewire function of igraph
}
