% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBINWeighted.R
\name{rewireWeight}
\alias{rewireWeight}
\title{rewireWeight}
\usage{
rewireWeight(data, number, rewire.w.type = "Rewire", dist = "Other")
}
\arguments{
\item{data}{The output of prepGraph}

\item{number}{Number of rewiring trials to perform.}

\item{rewire.w.type}{method to rewire weighted graphs}

\item{dist}{distribution for Garlashelli method}
}
\description{
makes the rewire for weighted networks
}
\keyword{internal}
