% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{robinFDATest}
\alias{robinFDATest}
\title{robinFDATest}
\usage{
robinFDATest(graph, model1, model2, legend = c("real data",
  "null model"))
}
\arguments{
\item{graph}{The output of prepGraph.}

\item{model1}{The viMean output of the robinProc function (or the viMean1 
output of the comparison function).}

\item{model2}{The viMeanRandom output of the robinProc function (or the 
viMean2 output of the comparison function).}

\item{legend}{The legend for the graph. The default is c("real data", 
"null model").}
}
\value{
Two plots.
}
\description{
The function implements the Interval Testing Procedure for 
testing the difference between the two curves.
}
\examples{
robinFDATest(graph=graph, model1=Proc$viMean,
model2=Proc$viMeanRandom)
}
