% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{robinGPTest}
\alias{robinGPTest}
\title{robinGPTest}
\usage{
robinGPTest(ratio)
}
\arguments{
\item{ratio}{The ratios output of the robinProc function (or the ratios1vs2 
output of the comparison function).}
}
\value{
The Bayes factor
}
\description{
This function makes a test between the curves, calculating the 
Bayes factor.
}
\examples{
robinGPTest(ratio=Proc$ratios)
}
