% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROBIN.R
\name{robinGPTest}
\alias{robinGPTest}
\title{robinGPTest}
\usage{
robinGPTest(ratio)
}
\arguments{
\item{ratio}{The ratios output of the robinRobust function (or the ratios1vs2 
output of the comparison function).}
}
\value{
A numeric value, the Bayes factor
}
\description{
This function makes a test between the curves, calculating the 
Bayes factor.
}
\examples{
my_file <- system.file("example/football.gml", package="robin")
graph <- prepGraph(file=my_file, file.format="gml")
graphRandom <- random(graph=graph)
Proc <- robinRobust(graph=graph, graphRandom=graphRandom, 
method="louvain", measure="vi",type="independent")
robinGPTest(ratio=Proc$ratios)
}
